Adding the code that mprotects the data segment, registers the sigsegv handler, catches the faults, sets up a seperate mmaped region to hold a bitmap of dirty pages and then updates the bitmap appropriately within the signal handler and fixes the permissions on the offending page.
This committed code needs more work. But for the most part it does what it
should.  Problems are:
	* at every ckpt, we mmap a new segment to hold our bitmap information,
	we never unmap this segment, it should be easy, but doing so causes
	problems
	* Currently I'm testing using a little program that manipulates the heap
	and ckpts periodically.  It works great until we do the increase the data
	segment test: We're suffering an internal segfault error.  This is where
	we segfault twice on the same page.  The second time we check our bitmap
	and discover that we've already faulted on this page, this indicates real
	problems so we Suicide.  Except that I happen to know that the test
	program is not causing any legitimate segfaults on its own.

By the way, for the time being we're only doing incremental checkpointing on the
data segment because it is usually about 99% of the total pages of the job.
But the logic and the structure is in place to extend incr. ckpting to the
other segments as well.