fixed the internal seg problem. The problem was that after ckpting, the old bitmap was still temporarily being used when the image was mprotected again.
There are a couple of solutions.  What I did was turn off the mprotect at
checkpoint time.  This is also nice because anything that is dirtied during
the checkpoint code is not dirtied by the user code and isn't necessary for
restart.  Except this may be problematic because condor may need some of its
own variables updated, such as condor_num_restarts, etc.

So the real best solution is to make sure that the bitmap and the current
protection on the image are the same.