Big changes.
+ Now, when we output the senders.C file, instead of it being case sections
in a big switch in a single function called REMOTE_syscall, it is now
seperate functions whose name is derived from the service the sender is
supposed to perform:

CONDOR_stat

now has a specific sender:

REMOTE_CONDOR_stat();

that MUST take the correct parameters for the stat call.

+ Changed the switches.C output to contain calls to the name mangled sender
calls specific to the service requested instead of to REMOTE_syscall.

+ The reason for all of this stuff is because gcc 2.96(a broken compiler, but
not is this context) really hated to take off enums from a va_list with va_arg.
The reason is because an enum can have an arbitrary type size and things on
a function argument stack are always promoted to an int sized quantity if
smaller. It is illegal to take of a quantity less than an int size from the
argument stack.

The way we used to fix this was we'd force the enum to hold an int
sized quantity thereby being implicitly type correct. Well, compilers
got around to finally calling what we were doing an error(legitametly)
and now forbids this trick.

Needless to say, this has caused a rewrite of a lot of code. On the plus side,
I've found a few run-time errors where arguments weren't being passed
correctly(through the va_list) and a few other oddities. This is good because
what used to be a runtime error is now a compile-time error.

+ fixed a bug where not supported calls wouldn't get emitted for switches
or senders. They are supposed to be emmitted because the shadow determined if
it should handle these calls.

+ Any code dealing with the receivers.C file has not changed.

+ As a side effect, you may not produce ONLY a switch from syscall.tmpl because
it will produce a call to the sender, which must exist at compile time. Of
course if you write your own in switches.special.C then you can do what you
want.