+ Fixed bug where job attribute string values with $$() macros did not get expanded if the string also contained quotation marks.
Some bug fixes in how the schedd handles globus universe jobs
with $$() macros :

+ Fixed memory leak: in globus universe, matched "machine" ads are now
deallocated once all $$() match values are stored in the job ad, or
the job ad is removed from the queue.

+ First time GetJobAd is called to expand the job ad, store
matched attribute *persistently* to disk w/ SetAttribute.  This way
if the schedd crashes without compressing the jobqueue.log, we are
still safe.