Added support for a submit_event_user_notes command in submit files.  It works exactly like the existing submit_event_notes command (used by DAGMan), but the two can be used together to provide not one, but *two* lines of arbitrary text inside a submit event.
Yes, implementing this as another parallel instance of the same code is
ugly, dumb, and redundant, but Alain and I both agreed it's the best of a
few bad options to quickly provide what EDG (specifically, Francesco
Giacomini's Planner Helper) needs: the ability to add a second notes line
to the submit event so that they can associate condor job ids with EDG job
ids, just as we use the existing one to associate condor job ids with
DAGMan job names.

We considered instead just adding support for appending more text to an
already-defined submit_event_notes line (e.g., "notes = $(notes) foo"), but
that would require a newline to be stuffed in, which introduces its own
problems, and would likely break XML logs.  (An as aside, Alain may add the
"attrib = $(attrib) foo" feature for submit files anyway, but not to solve
this problem.)

Long term, I think the right solution is to allow a classad to be tacked
onto any log event, into which DAGMan, condor, or the user can stuff
arbitrary attrib/value pairs.  Or, in a classad-log world, maybe just allow
them to insert the attrib/values pairs in the event's classad.

Anyway, if/when we do implement this right, backwards-compatibility with
this hack may suck (or not be worth doing, which sucks in its own way), but
we think that problem already exists with the current lognotes feature, and
adding a second one doesn't fundamentally change it for the worse...

Mea culpa.