Initial checkin of soap and built-in web server support.
DaemonCore now examines if the incoming connection to the daemonCore
command socket is HTTP.  If so, and it is a POST command, it is
dispatched as a soap rpc.  If it is a GET command, it is a web
page request (mini web server).  If it is CEDAR, it is treated as usual.
The collector request queries have been soaped.
Much more work to do.