First pass at the schedd changes needed to support Jaime's grand vision for mirroring jobs/CondorC.  I just work here.  :). Specifically:
+ Added support for WantMatching attribute in the job ad, which if present,
controls if the schedd will try and negotiate a match for this job or not.

+ Added ScheddBday attribute into the job ad.

+ If MirrorSchedd is defined, start a gridmanager per Owner,MirrorSchedd
pair. Use meta-operators in constraint given to gridmanager.

+ On schedd restart, if job has gone over to mirror mode
(MirrorActive = True), schedd shouldn't twiddle job status, etc. It should
start a gridmanager (not optional).

+ On hold/remove/release, treat like globus universe, plus whatever should
be done for a shadow. On completion, send a remove signal to the
gridmanager.