Ok, again lots of changes. Hopefully there will be smaller and smaller updates as time goes on.
So, I changed all the XAndY structs to contain other structs instead
of pointers to structs. There was little point in using
pointers. Maybe for efficiency later, but that's later.

Also, added some code to soap_helpers.cpp to convert ClassAdStructs to
ClassAds, and added a simple implementation of
discoverJobRequirements. discoverJobRequirements was tested with an
Axis client, so you can write Java to talk to the soapy schedd and it
isn't TOO ugly using Holders.

I would like very much to be able to test all the old functionality,
but this means rewriting all the tests, which I should point out to
Swaroop. Actually the rewrite should be easy since it is just method
signatures that changes a bit, not really the semantics.