Ok, changed the discoverJobRequirements() to use a FileTransfer object instead of duplicating code. To do this I created a LooseFileTransfer object (loose because, well, it gives you access to more than the "chaste" FileTransfer object) that provides access to the InputFiles member variable of the FileTransfer object. To do this I had to make FileTransfer::InputFiles protected instead of private. I also ran into problems with List not allowing copy constructors or assignment constructors, which took me too much time to figure out -- the compiler was not giving errors at first. Then I wasted a good hour/hour and a half after everything SHOULD have been working because file_transfer.o was not recompiled, which means the "this" of the LooseFileTransfer object had a SLIGHTLY different structure from the "this" of the FileTransfer object -- talk about a really annoying thing to waste an hour on!
Anyway, to get this all compiling I also modified Imakefile to specify
loose_file_transfer.o as a requirement for linking condor_schedd.

So, this is the first somewhat major (as in something that is how is
should be) commit that I can make, and thankfully it is also a small
set of slightly changed files, i.e. not huge and nasty.