README: updated documentation
history++.C: added user=quill to the dbconnection string. it is purposely not appended to the string.  this is just done as a sanity check in the clients but in the case of the quill daemon, putting this in the end might result in an overwrite due to strcpy (see daemon documentation in JobQueueDBManager::config function.

jobqueuedbmanager.C: added user=quill to the dbconnection string. it is purposely not appended to the string.  putting this in the end might result in an overwrite due to strcpy (see daemon documentation in JobQueueDBManager::config function.

prober.C: cosmetic changes