modified these three files to deal with various memory leaks.  Not all leaks are resolved.
Most of the leaks in the client (history++.C and historysnapshot.C) were due to calling param() and not freeing the resulting string.

most of the leaks in jobqueuedbmanager.C was from not calling releaseQueryResult.  There still is a recurring 112 byte leak on every polling period.
Purify points this to someplace in the postgres code (PQmakePGemptyResult) but I think (actually I hope) its someplace in my code.