Made extensive changes to the quill server incorporating all the changes suggested in the various code review sessions.  These changes include but are not exclusive to:
1) Minor changes involving renaming of variable names and ATTR_
instead of hard-coding them wherever they're being put inside classads
sent to the collector

2) Modifying them suitably to recognize the fact that quill daemon
now sends QUILL_AD and not a fake SCHEDD_AD to the collector.  This
involved changing the fail-over algorithm which is now documented in a
WISDOM file inside condor_q.V6 directory and will in turn be added to
the manual

3) Modifying sql statements: a) to be idempotent and b) added a
EnteredHistoryTable attribute to the history_horizontal table

4) removing the reading of xactState variable.  It is still around
and updated to reflect the xact state of the schedd but it is never
checked.  Kept only for future potential use.