+ Fixed the quill interface for condor_q to understand that you can have this type of a command line:
condor_q 13.4 54 567.7

and have it return everything requested.

+ If you query the quill daemon, then you end up pulling all of the ads
out, but if you query the db, then all of the specific jobids are constrained
properly in the sql statments. This will be fixed in a future revision,
hopefully, as it will increase performance.

+ The behavior is correct in the case of the query to the quill daemon, only
that the condor_q does the filtering instead of the sql server.