Foremost, we now have requirements evaluation for local jobs. I editted condor_submit so that local|scheduler universe jobs no longer have the default requirements added to their job ad.
This commit contains the bulk of my cron job code (currently disabled) and the job deferral failure handling

For the job deferral, I added a case statement in child_exit() so that if the job misses it deferral time, we will always kick it out with an error message in the logfile

There are two new data structures that we use to keep track of who needs a cron schedule and who doesn't. If a job is in the cronTabExclude list then we know that it never will need a schedule and we can always skip it. For others, we store the cached crontab object in the cronTabs hashtable

There is a new method to walk through the job queue and calculate all the cron times as necessary. We also will try to calculate the cron time when a job enters the queue or comes back after executing