Added new configuration parameter: # Enable/disable job history file logging to the file, # $(STORK_LOG_BASE).history.   If enabled, the job history file # provides input for the stork_status completed job query command, and # initializes the first job id upon server startup.  Currently, this # stream is not truncated, or rotated.  Default = true.
This implementation still creates a history file, and the ClassAd_Reader class
still creates this file, albeit zero size.  It would be better to disable
history logging by setting a null history filename, and not operate on the
history file at all.  However, the ClassAd_Reader class always assumes the
history file exists and is open, for all methods.  It's too risky to modify
several methods in the class one day before the SC05 show.