A big overhaul on the internal mechanics of the Starter For all the methods that are registered with the daemon core there are two versions: a "remote" and "regular" method. The daemon core will call the "Remote" version of the method (RemoteHold instead of Hold). The remote version will notify the JIC about the command, and then just call the regular version of the method which does all the same work as before. This is necessary so that I can remove|hold jobs in the JIC without the Starter thinking the signal came from the outside.
Added the "suspended" flag to the Starter that fixes when a Suspend call arrives and a job is currently
being deferred that when the deferral timer fires off the job will be suspended. The old behavioir was
that if a deferred job gets suspended the job is just removed and the Starter exits.