Previously, the global Q_SOCK used to store just a ReliSock* that connected to a client.
Now Q_SOCK is a new object type (QmgmtPeer) that store *all state*
for a remote qmgmt client.  Not just a possible ReliSock*, but the
transaction state, a bunch of qmgmt state (like the active cluster num, etc),
who the peer authenticated as, etc.

This change was made to allow more than one client/transaction to
be processed at once, as required for SOAP clients.