Use the quill++ schema for reading from the job queue.
This is structurally close to the way Quill did it, but intead
of num and str tables, theres hor and ver tables for clusters
and processes. Basically, all str tables got converted to hor tables
and num tables were treated as ver.

The signature for getJobQueueDB changed to take a scheddname
instead of an owner name - we never used owner in quill anyway.
I changed the order to make it clear to callers that it's
different (the compiler will enforce the types)

todo - update historysnapshot