+ Removed some unnecessary cruft from the protocols between the agents. Everything is now classads (except the encapsulation method string).
+ Added a registration/reaper callback the schedd may specify to be
called on a per transferd basis by the TDMan object.

+ Added default reapers for everything. Currently they don't do anything
beyond the barest necessities.

+ Now the usual behaviour is that any callback may NOT assume control
of the passed in pointers UNLESS TREQ_ACTION_FORGET is used as a return
value.  Ultimately, I should convert this callback system to useing
const references to the transfer request objects, so we'll never have
to worry about memory allocation again concerning transfer requests
(and potentially transfer daemons).

+ Fixed up some comments

+ Added a condor wide daemon exit code to detect when a child has excepted.

+ Added a simple timeout feature where if the transferd has an empty transfer
request list for more than 20 minutes, it exits. This information is
passed on the command line and currently cannot be changed once set. Obviously,
that'll change in the future sometime.