+ Added a mechanism for condor_submit and condor_transfer_data to use different methods of reading/writing the spool files into Condor.
+ One method is stm_use_schedd_only, which is the original method of
reading/writing spool files into Condor. The original method was that the
schedd is the sole point of contact, and it would fork a threaded file
transfer object to handle the request.

+ Another method is stm_use_transferd. In this method, the client contacts
the schedd for a location and idenfitication string of a transferd, then
it contacts the transferd to perform the read/write.

+ As of this checkin. condor_submit and condor_transfer_data default to
use the original schedd only method.

+ Added -stm to condor_submit and condor_transfer_data which specifies what
sandbox transfer mechanism to use.