Added AmazonUserData attribute to an Amazon job.
The attribute can only be added to a job via +AmazonUserData="..." in
a submit file. condor_submit does not handle the attribute
directly. The value of the attribute must be base64 encoded,
e.g. "echo -n 'MyUserData' | openssh base64"'s output. The value of
the attribute is passed from the gridmanager to the amazon gahp, then
to the condor_amazon.pl script via a -u argument, and finally to
EC2. This code has been tested with and without +AmazonUserData in a
submit file, but corner cases such as very large user data has not
been tested.