Prevent the gridmanager from creating temporary security groups, which it does not properly clean up on recovery.
This patch mostly comments out portions of the code that would create
a temporary security group if one was not specified in a job's submit
file. It also makes sure that if a security group is specified in a
job's submit file that it is used. The testing of this patch consisted
of both not specifying a security group at all, which meant the EC2
user's default security group was used, and specifying a security
group, which meant it must be used. Both test cases succeeded. In
neither case did the gridmanager try to, inappropriately, delete a
security group it did not create itself.

While writing this code it became clear that a failure in the
gridmanager can result in GM_DESTROY_SG seeing an empty
m_groups_names, because it is only setup during the GM_CREATE_SG, and
as a result leaking security groups into EC2.