Working gSOAP amazon-gahp_worker_thread
Discovered the previous compilation and segfault issues were likely
not because of the gSOAP version but because C code was being
generated instead of C++. This patch uses C++ everywhere, and includes
many changes to gsoap_commands.C to support the different structures
in ec2Stub.h.

Also, SSL authentication appears to be failing so the AWS_URL is
changed to http from https.

The only function that has been tested is AMAZON_VM_KEYPAIR_NAMES, but
it worked.