Remove temporary deadlock between amazon gahp and gridmanager on startup
When the Amazon gahp first starts up, it makes a blocking connection to
the gridmanager to send a CHILDALIVE message. This usually happens when
the gridmanager is blocked waiting for a response to a gahp command.
After 20 seconds, the CHILDALIVE times out and the deadlock ends.

By unsetting CONDOR_INHERIT in the environment on startup, we prevent
this temporary deadlock from occurring. The Amazon gahp io process
doesn't know it has a daemoncore parent, but that's ok. It knows when to
exit from stdin/out closing.