Handle empty results for AMAZON_VM_STATUS gahp command
The VM status call to EC2 can return nothing if the given instance has
been terminated for long enough. Now, the Amazon gahp propagates this
result rather than generating an uninformative error.
The gridmanager is now able to handle this new behavior from the gahp
command.
And AmazonJob will treat the job as 'terminated' if it sees this case.