implemented procd_snapshot
The procd_snapshot tool queries ProcAPI for information about all
processes on the system and dumps this information to stdout as a
series of procInfo structures. The helper tool procd_snapshot_format
can be used to read a series of binary procInfo structures from
stdin and display them in human-readable form to stdout.