Added two new targets to each Windows project: one for VS and one VS Express.
This allows people without a huge budget, or interest in future development
to use the free version of Visual Studio to build Condor on their machine.

The new targets store all their output in the same place that the original
targets do, so no changes to the distribution and packaging scripts were
needed.