Each node has three lists which store the parents, waiting list, and children for a node. I converted them to STL sets instead of the CTL SimpleLists they were for performance reasons.
In a dag where there is one parent, a million children who are all
siblings, and one end node, the parent and end nodes have 1 million
entries in their children and parents sets respectively (a common dag
geometry among our customers). As new parents or children get added,
they are looked up in the lists to see if it already exists since it
is a violation to have two or more edged from one node to another. This
look up was happening in linear time, for each added dependency, which
has ugly complexity behanvor.

So, the edges were being treated as sets, so I changed the data structure to
be an actual set. The complexity behavior is much better for the set and
a parse time of 2 hours and 30 minutes on a 3.2GHz machine with 4 GB of ram
changed to 2 minutes and 30 seconds. I could probably do better with some
more rearranging, but a 60 fold improvement is good enough for now.