Fixed spurious spool dir creation and deletion.
In the past, we have created and deleted sandboxes for
submissions that do not require file transfer.  This
is important in environments using NFS spool directories
(or spool directories on other shared filesystems),
since these operations are between one and two orders
of magnitude slower on NFS (and may also unacceptably
load the NFS server).

We ensure that job sandboxes are deleted for jobs whose
CleanUpSpuriousJobSandboxes attribute is set.  This
attribute was added to work around leaking job sandbox
directories for jobs that were running before upgrading
to a version of the schedd that does not create spurious
job sandbox dirs.

This is a fix for ticket #749 and RHT BZ 523806

RELEASE NOTES:

To avoid leaking directories, run this command before
upgrading to this version:

condor_qedit -const 'toLower(ShouldTransferFiles)=?="no"' CleanUpSpuriousJobSandboxes ALWAYS