This checkin makes it so that making the individual stduniv test programs will produce symlinks to the condor_tests/dmtcp-local directory so the tests have available the binaries and libraries for DMTCP.
There is also a *vile* hack in this checkin.

Due to issues described in #974, specifically the one where the nmi test glue
makes the test suite radically differently than a developer would, I altered
the target for all_symlink_target() to be a double-colon target.

Then I added a dependancy for batch_test.pl to also build DMTCP and the
associated condor glue library to DMTCP. While the dependency to build
DMTCP is in the all:: target, the nmi test glue as of this time doesn't use
that target.

This had to happen because otherwise I would have had to change the glue, and
that is exceedingly difficult due to the many reasons of #974.

This is for ticket #366.