Added the .run file test harness to be used for low level testing of DMTCP and any other checkpointing system. These tests run outside of Condor. The Condor.pm file cannot be used to implement these tests since by definition they are not stduniv so the user event log will hold very minimal information and we'd be tied to having already integrated the checkpointer into Condor in order to understand its output from Condor itself.
The StaCkpt/DMTCP.pm files is a module to be used by the .run test program
in our test suite. Said module implements a master/slave relationship
between two processes. The slave runs the dmtcp coordinator, the test
job under the dmtcp checkpoint program, and the dmtcp command tool to
interact with dmtcp (as to ask it to checkpoint, etc).

This design is the result of a series of iterations of design and code where
when dmtcp died or it or the test deadlocked we'd get a -32 timeout in NMI
and that is worse than useless. The asynchronous I/O model this codebase
embodies allows us to discover the error and report what it is when
we end up in a state we don't expect. This is far more useful to us.

This is for ticket #366.