Add option of claiming paired slots to REQUEST_CLAIM. #3931
The schedd can add special attribute _condor_SEND_PAIRED_SLOT to the
request ad to indicate it wants to claim a paired slot's partner. In
this case, the startd can respond with a new reply code, followed by the
partner's ad and claim id. Claiming of the partner can be disabled via
the new configuration parameter CLAIM_PAIRED_SLOT.
This commit only implements the client side of this protocol change.