Swap claim to paired slot when job enters output stage. #4291
In the schedd, claim the partner of a paired slot. When a job indicates
that it is in output phase (by setting JobTransferringOutput to True in
the job ad), the schedd will attempt to swap the claim with the paired
claim, allowing a new job to run on the original slot.