Enable larger UDP (safesock) fragmentation sizes. #4321
Add knobs UDP_LOOPBACK_FRAGMENT_SIZE used when going to the loopback
interface (default to largest message size, currently 60k), and
UDP_NETWORK_FRAGMENT_SIZE when going to any non-loopback address
(default to 1000 bytes as before). Note these are not in the param
table because the defaults are based upon constants in CEDAR not
available in the param table code.