Prefer module's copy of symbols over daemon's.  #3718
This changes the module loading flags to dlopen to use RTLD_DEEPBIND.

We use "deep bind" here to make sure the library uses its own version of the
symbol in case of duplicates.

In particular, we were observing crashes in condor_shadow (which statically
links libcondor_utils) and libclassad_python_user.so (which dynamically links
libcondor_utils).  This causes the finalizers for global objects to be called
twice - once for the module version (which pointed at the shadow's memory) and
again at exit().

With DEEPBIND, the finalizer for the dynamic library points at its own copy
of the global and doesn't touch the shadow's.  See #3718