condor_tests: handle job_test_scheddroation abort
As per the new improvement when the submitted jobs are timed
out we will purge all the jobs from the queue. We do not want
to end the test case with failure when the jobs are purged
so we add a abort handler, but the name of the handler used
for aborting was incorrect in the earlier commit, fix it.